annoying warnings

+++ include/lib_formatter.hpp
@@ -79,76 +79,76 @@ namespace tools
             int16_t exponent = 0;
 
             if (value >= FP::posExpThreshold) {
-                if (value >= 1e256) {
+                if (double(value) >= 1e256) {
                     value /= 1e256;
                     exponent += 256;
                 }
-                if (value >= 1e128) {
+                if ((double)value >= 1e128) {
                     value /= 1e128;
                     exponent += 128;
                 }
-                if (value >= 1e64) {
+                if ((double)value >= 1e64) {
                     value /= 1e64;
                     exponent += 64;
                 }
-                if (value >= 1e32) {
+                if ((double)value >= 1e32) {
                     value /= 1e32;
                     exponent += 32;
                 }
-                if (value >= 1e16) {
+                if ((double)value >= 1e16) {
                     value /= 1e16;
                     exponent += 16;
                 }
-                if (value >= 1e8) {
+                if ((double)value >= 1e8) {
                     value /= 1e8;
                     exponent += 8;
                 }
-                if (value >= 1e4) {
+                if ((double)value >= 1e4) {
                     value /= 1e4;
                     exponent += 4;
                 }
-                if (value >= 1e2) {
+                if ((double)value >= 1e2) {
                     value /= 1e2;
                     exponent += 2;
                 }
-                if (value >= 1e1) {
+                if ((double)value >= 1e1) {
                     value /= 1e1;
                     exponent += 1;
                 }
             }else if (value > 0 && value <= FP::negExpThreshold) {
-                if (value < 1e-255) {
+                if ((double)value < 1e-255) {
                     value *= 1e256;
                     exponent -= 256;
                 }
-                if (value < 1e-127) {
+                if ((double)value < 1e-127) {
                     value *= 1e128;
                     exponent -= 128;
                 }
-                if (value < 1e-63) {
+                if ((double)value < 1e-63) {
                     value *= 1e64;
                     exponent -= 64;
                 }
-                if (value < 1e-31) {
+                if ((double)value < 1e-31) {
                     value *= 1e32;
                     exponent -= 32;
                 }
-                if (value < 1e-15) {
+                if ((double)value < 1e-15) {
                     value *= 1e16;
                     exponent -= 16;
                 }
-                if (value < 1e-7) {
+                if ((double)value < 1e-7) {
                     value *= 1e8;
                     exponent -= 8;
                 }
-                if (value < 1e-3) {
+                if ((double)value < 1e-3) {
                     value *= 1e4;
                     exponent -= 4;
                 }
-                if (value < 1e-1) {
+                if ((double)value < 1e-1) {
                     value *= 1e2;
                     exponent -= 2;
                 }
-                if (value < 1e0) {
+                if ((double)value < 1e0) {
                     value *= 1e1;
                     exponent -= 1;
                 }
@@ -162,7 +162,7 @@ namespace tools
             exponent = normalizeFloat(value);
 
             integralPart = (uint32_t)value;
-            double remainder = value - integralPart;
+            double remainder = (double)value - integralPart;
 
             remainder *= 1e9;
             decimalPart = (uint32_t)remainder;
